The Submit button on activity_number_output.xml now shifts the fragment from fragment_results.xml to fragment.submit.xml.
The new fragment still displays the largestInt variable but now gives the player an option to submit their name and score to FragmentScoreboard.

The ability to submit the score actually doesn't exist yet. The next iteration will be focused on using an ArrayList and a for of data persistence to accomplish this task.

Another big change in this commit is changing the bundle in NumberOutput.java into a class variable and initializing it in our initialize method with the rest of the variables. This improves the re-usability of our largestInt variable and bundle without having to continually re-instantiate it.