Functionally, not many changes going on here. Mostly worked on structural refactoring. Unfortunately this has been somewhat devastating to the UI, but I will worry about that later. Data persistence has yet to be implemented.
NumberOutput.java has completely been changed and now relies specifically on the fragments it houses. All of the functionality that was once located their has been moved either to FragmentResults.java, FragmentSubmit.java or FragmentCommitResults.java.

The big changes here come from FragmentCommitResults.java. It now not only has the capabilities of altering the results fragment but also itself. I think the interface once completed will be a bit cleaner, but this done limit the number of options available to the user. Not like this application does a whole lot though.

Additionally a lot of comments, unused imports, and logs were removed. Just some general housekeeping.